KVM test: minor pci_hotplug fixes
- Put the PCI device removal code in a finally clause.
- Use kvm_vm.get_image_filename() instead of os.path.join().
It's a bit cleaner because if we ever change the names of image parameters
we'll only have to change the code in one place.
Also, the way os.path.join() was used lead to image filenames being prefixed
with 'images/' twice, e.g. 'images/images/foo.qcow2'.
- Make some failure messages clearer.
- Remove 'only Fedora Ubuntu Windows' from the fmt_raw variant.
'only' works for things that have already been defined, but the guests are
defined later.
- Remove unused 'modprobe_acpiphp' parameter.
- Change 'online disk' to 'online' in pci_test_cmd for Windows ('online disk'
doesn't seem to work).
- Remove the unneeded telnet/ssh/guest_port parameters from the Windows
block_hotplug parameters.

Changes from previous version:
- When deleting the PCI device compare the output of 'info pci' rather than the
output of the pci_del command.
- If an exception is raised in the 'try' block, do not raise an exception in
the device deletion code, because it will mask the first exception.
- After pci_del, give the device some time to disappear from 'info pci'.

Signed-off-by: Michael Goldish <mgoldish@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@3938 592f7852-d20e-0410-864c-8624ca9c26a4