Control file for KVM kernel module load/unload
This is the first of several steps that will add autotest control files
that are tailored to KVM testing.  In these tests we simply load the
kvm module, run existing tests, then unload the kvm module.  The
purpose of these tests is to show that there is no negative impact to
a machine when the KVM modules are loaded.

Signed-off-by: Anthony Vallone <avallone@google.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@3978 592f7852-d20e-0410-864c-8624ca9c26a4