Adds support to wait_down and wait_for_restart to watch for changed boot_id values (using /proc/sys/kernel/random/boot_id). This avoids a race condition where if a machine was able to successfully shutdown and boot up again after you started a reboot and before we started polling for the shutdown, it would just appear that the machine failed to shutdown when told.
As a result the Host.wait_down method now treats the case where a
machine is NOT down but has a new boot_id as being "down", because
this implies that the machine did shut down (and subsequently came
back up). This does mean that you cannot assume that a successful
wait_down implies that the machine is now down, but that was never
true anyway.

Risk: High
Visibility: We can now reliably detect the restart of machines with
    very fast restart times.

Signed-off-by: John Admanski <jadmanski@google.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4096 592f7852-d20e-0410-864c-8624ca9c26a4