Rearrange the why isn't my job running logic so that it checks all of the preconditions first before looking at the Ready state of the machines.
Also have it list all reasons the job is not running rather
than only the first reason and dying.  That way when someone uses it
they don't think "oh, its just that, let me fix that" only to have it
not work and have to rerun it again later to see the next reason.

Signed-off-by: Gregory Smith <gps@google.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4160 592f7852-d20e-0410-864c-8624ca9c26a4