KVM test: Removing qemu/qemu-kvm variants from tests
The idea of having a qemu/qemu-kvm variants block
turned out to be a bad idea, since we are never meant
to run them in combination. So let's roll back
this change and give some examples on tests.cfg.

Also, cleaned up tests.cfg.sample, and introduced
Fedora 12 as the default host, as it is the latest
Fedora upstream.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4188 592f7852-d20e-0410-864c-8624ca9c26a4