KVM test: Define defaults for KVM test
Define reasonable defaults for the kvm test, to make it
easier to start using KVM autotest and encourage people
to play with it. The defaults are:

 * qemu and qemu-kvm are expected to be found under
   /usr/bin.
 * The auxiliary directories are expected to be
   found under /tmp/kvm_autotest_root, since some
   people might want to use a nfs share to provide
   cds.

A first time config script will be provided to assist
people creating their first kvm test setup so it can
be done answering a few questions and leaving the
heavy lifting to the script.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4189 592f7852-d20e-0410-864c-8624ca9c26a4