KVM test: Add a get_started.py script
In order to make it easier to get started with KVM
testing, add a script that helps with setting up
directories, paths and iso files for the default
test configuration (Fedora 12 guest install, boot
test, shutdown). Once the first idea of how things
work is formed, one may start to tinker with
configuration files.

Added pointers to the online documentation.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4192 592f7852-d20e-0410-864c-8624ca9c26a4