New, more consistent way of representing relationships in RESTful interface.  A relationship between two entrys is represented as a relationship itself, contained in a collection.  For example,
resources/labelings?host=myhost

retrieves all labelings for host myhost, each with a URI like

resources/labelings/myhost,mylabel

which represents that the label mylabel is applied to the host myhost.  The user can POST to the former URI to relate two entries and DELETE the latter URI to break the relationship.  This makes it much more consistent with the rest of the interface.

There are various other refactorings included in here that came up as I worked through this change to make things cleaner.

Signed-off-by: Steve Howard <showard@google.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4246 592f7852-d20e-0410-864c-8624ca9c26a4