Split netperf/netpipe control files out into tests
From:kdlucas@google.com

So I've split the netpipe and netperf control files into separate control files and a .py file for each one, so that it will be easier to add additional tests that use the logic of the python scripts without duplicating code.

I'm also taking advantage of the new added function to get the platform_label, as in my testbed I have assigned a net_server platform label for the machines I want to use as the network side of these tests.

So, the attached patch now adds two methods to client/common_lib/hosts/base_classes.py, and it stores the current state so when you enable IPFiltering it will restore to it's previous state.



git-svn-id: http://test.kernel.org/svn/autotest/trunk@4288 592f7852-d20e-0410-864c-8624ca9c26a4