Get the bma150 autotest working.
Explicitly report module unloaded if the open of /dev/bma150 fails.
This test only works if the bma150.ko kernel module is loaded.  And that cannot
currently be done on any of our systems.

But, when we can get to the device, we have a test for it.

Review URL: http://codereview.chromium.org/1164007