Add a test for the tsl2563 series of light sensor chips.
We do not currently have I2C access to one, but once we do we
can point this test at it.  hardware_tsl2563.py may need to be
modified to point at the right sysfs file with the -f option,
or the default in tsl2563tst.c can be modified.

Review URL: http://codereview.chromium.org/1389001