Add keyboard manufacturing/factory test.
Initial commit of files.  Description of test follows.

Intended for use during manufacturing to validate that all keyboard
keys function properly.  Run normally, this program will display a
keyboard image and keys will be highlighted as they are pressed and
released.  Once all keys have been used, a brief 'PASS' message will
be displayed and the test will terminate.  After the first key is
hit, a countdown will begin.  If not all keys are used in time, the
test will fail with an 'ERROR' message that is displayed forever.
There is also a setup mode, to facilitate programming the expected
key bindings, described in the KeyboardTestSetup class.

Review URL: http://codereview.chromium.org/1519006