This patch does the following:
1) In older versions of sar, specifying an interval and
a count of 0 means that reports will be generated
continuously. For newer versions, just ommiting count
does it, a count of 0 is invalid. Handle this difference
testing the command during the initialize() method.

2) Try to terminate the process using the subprocess
method terminate() (which only sends a sigterm to the
process and does not raise exceptions if the process
was already terminated).

3) Add some docstring documenation for the profiler.

Signed-off-by: Feng Yang <fyang@redhat.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4370 592f7852-d20e-0410-864c-8624ca9c26a4