switch to autox.py and robustify login/logout code
This CL switches from using the autox binary to using autox.py.  This includes backwards
incompatible changes to site_login.py, and I haven't had a chance to fix the affected callsites
yet.

I've also made the login/logout code a little more robust.  Now it'll make sure that the login
manager is running under a NEW pid, that X is running, and that at a window is visible before
assuming we're ready to log in.

All of the wait loops have been refactored into wait_for(...), which spits out logging.info()
messages that could be parsed later to determine how long the operations are actually taking.
Perhaps this could make it into a different, non-login specific library soonish.

I'm out of the office for the next few days, but wanted to get this out there before the
trunk totally passed it by.  I'll return on Wednesday to finish the job.

Review URL: http://codereview.chromium.org/1534001