Hackbench test: Fix performance keyval generation
The hackbench test wasn't generating performance keyvals,
due to some incorrect assumptions at the postprocessing
code (the output for the test also has a debug line with
the number of tasks spawned, so the output generated will
not start with 'Time:', as the code assumes). Let's fix
that problem.

Also, in order for people to apply their custom
postprocessing to results (they might want to test out
some 3rd party postprocessing script, for example), is
allways good to keep record of the raw output generated
by the test, per iteration. I plan on doing the same
for the other benchmarks present on autotest.

Updates from v1:

 * Updated the test to postprocess keyvals per iteration

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4410 592f7852-d20e-0410-864c-8624ca9c26a4