KVM Test: Fix bug that check_image.py will check removed image
If set remove_image to yes, the image will be removed before
check_image.py is executed as post command.  But check_img.py
still try to check the removed image When it is running.
This patch update check_image.py and let it do not check
non-exist image.

BTW, could we first run post command then remove the image?

Signed-off-by: Feng Yang <fyang@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4412 592f7852-d20e-0410-864c-8624ca9c26a4