KVM test: Introducing Opensuse and SLES unattended installs
Adds support to do unattended install of Opensuse11, patch adds
new autoyast files (OpenSUSE.xml and SLES-11.xml) and creates
entries for unattended install in tests_base.cfg. The patch also
introduces an optional parameter, post_install_delay in the config
file that allows people to adjust a wait time. In this particular
case, the timeout is for autoyast to finish configuration.

Signed-off-by: Yogananth Subramanian <anantyog@linux.vnet.ibm.com>
Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4428 592f7852-d20e-0410-864c-8624ca9c26a4