parallel.py: Fix the client test recycled pids problem
In a client control file calling a lot of times of job.run_test(), the
system's pid_max could be reached and the pids are recycled for the
tests run in child processes, causing the parent process mistakenly
determines the status of the child test.

This change fixes the problem by renaming the serialized
client/results/default/debug/error-pid file so that they are not seen by
the parent process after first use.

Signed-off-by: Jinxin Zheng <jzheng@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4501 592f7852-d20e-0410-864c-8624ca9c26a4