Land the first login test that uses a local mock of the google accounts server
With this CL, in our tests, authentication is no longer performed against
live servers.  Instead, we spin up a local DNS server that will lie
and say that all sites resolve to 127.0.0.1.  We use DBus to tell
flimflam to use this DNS server to resolve addresses.  We then spin up
a local httpd that will respond to queries at the Google Accounts
endpoints.  We clear the DNS setting and tear down these servers in
cleanup().  The caller can specify handlers for the three pertinent
endpoints, but default ones that return HTTP 200 exist and will be
used if nothing is passed in.

Review URL: http://codereview.chromium.org/1752022