Temporarily disable firmware_VbootCrypto autotest.
Until we figure out a way to run these in 64-bit mode, this autotest is useful and accurate as it is testing the reference code in 32-bit mode where as the code is run in a 64-bit environment. Disabling it will also stop the unnecessary build breakage and help keep build-waterfall-going-red rage down.

Review URL: http://codereview.chromium.org/2602001