KVM test: Improvements on build subtest
* Now koji based builds use the koji python API, so the
install operation is faster and takes up less space
* It is possible to specify commits for each repository on
the git install method. Also, you can specify a list of
patches to be applied to each repository as well
* New 'yum' install method, that allows one to specify
a list of kvm packages to install from
* It is now possible to save the result of the builds
(either source directories or rpms) on the test resultsdir
in the form src.tar.bz2, controlled by the parameter
'save_results'.
* Updated build.cfg.sample to document the new behavior
and options

Also, fixed a bug on the git repo checkout code, to
correctly point out that we can optionally get a
particular commit of that git repo.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4588 592f7852-d20e-0410-864c-8624ca9c26a4