Restores proper support for handling unexpected client terminations after the server_job.record refactoring.
The original record implementation just dumped whatever status logs
the client sent it into the status.log, so if the client died on you
there was never a problem with the client leaving the server-side
logging in a funky state. Now that client job records are being
properly routed though the server-side job.record a client death can
leave the server logging in a broken state.

This implementation adds a special get_record_context method which can
be used by the autotest.py code to retrieve the current job.record
context at the start of the client job, and then restore it at the end
so that the server side logging is put back into a known-good state.

It also removes a bunch of self._record_prefix uses in the server_job
code which were rendered useless by the job.record changes, but
weren't removed.

Signed-off-by: John Admanski <jadmanski@google.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4606 592f7852-d20e-0410-864c-8624ca9c26a4