Patches to turn off PIE on hardened for autotest. BUG=3981 TEST=unit,autotest
Various code in autotest is incompatible with PIE on x86 (it uses %ebx).  Since these are tests, it seems more apprpriate to turn off PIE, rather than to update the individual source trees with patches.  Unfortunately, not all of the code in autotest respects adding -nopie in the ebuild, so it requires patches and changes within the autotest tree itself.

Review URL: http://codereview.chromium.org/2719004