Change netperf wait time
This change allows site_wifitest to take control over the wait_time variable, eventually passed as the "-s" option to netperf from within the autotest system.  This parameter can be changed in the config file or the test control files, and is set to a default value lower than what is in netperf2.py, since with values larger than 4 it appears the remote gets board and quits on us.

BUG=
TEST=Reran full tests

Review URL: http://codereview.chromium.org/2806014