Revised login_LogoutProcessCleanup test
TEST=manually
BUG=none

Patch from James MacLean <wjmaclean@chromium.org>.

The file login_LogoutProcessCleanup.py file has been modified to:

1) add a process that ignores SIGTERM, in order to test that such processes
   are killed properly, and

2) before logout generate a list of 'chronos' processes in order that we can
   check to make sure they are all terminated properly during the logout.

Step 2 is necessary since not all 'chronos' processes are created by the
session_manager, and thus a test of descendancy from session_manager is
not sufficient to ensure all 'chronos' processes have in fact been killed.

This approach still allows for a race condition in which some 'chronos'
process is created between generating the list, and logging out. Such a
process, if it escaped being killed during logout, would constitute a
false positive for the test (it should have failed). However, this seems
to be the best we can do without inserting commands into the boot-script
structure, which is not feasible. (Specifically, a list could be compiled
during the time while the UI is not running, as controlled by
/etc/init/ui.conf, to verify that no 'chronos' processes exist at this
time.)

Review URL: http://codereview.chromium.org/2779003