Update ProcessPrivs tests, return login_RemoteLogin to BVT
In addition to adding login_RemoteLogin back to the BVT, as it passed the last 5 times it ran in the nightly.  It also passed when I run the BVT locally, and when I manually run it right after login_BadAuthentication (which used to make it fail).

Also, removing stale ProcessPrivs tests; the authors update their own control file, but not the BVT suite file, so we've been running a test that's expected to fail (ibus-X11) for some time now.

BUG=chromium-os:3997
TEST=ran the BVT locally, also watched login_RemoteLogin pass the nightly many times

Review URL: http://codereview.chromium.org/2818036