Make the test paramaters for ProcessPrivileges private to the test
This prevents duplicating a list of processes and their required
properties in two different places; the duplication was causing bugs
as sometimes only one list would be updated.

The code motion also uncovered dead code to be removed and other
opportunities for cleanup to be exploited.

BUG=4388
TEST=run each control file manually on an ST8; test error cases with temporary script change

Review URL: http://codereview.chromium.org/2853021