Move flashrom mapping selection scripts into autotest itself
This is a fix to http://codereview.chromium.org/2838020 .
Since we decided to use the 'packaged shell-balls' for firmware
update/setup, there will not be a ebuild package like 'firmware-utils'
to store the mapping selection scripts.

According to latest information from BIOS vendor, the mapping
scripts won't be easily changed for all x86 platforms, so we decide
to directly put the mapping selection scripts inside autotest.

NOTE: there are some magic numbers in the script. We believe this to be not confidential because they are common for all x86 platforms.

Review URL: http://codereview.chromium.org/2852040