Fixed SysStat init to find power supplies on non-ACPI systems.
Previously this test would look for power supplies of the format:
/sys/class/power_supply/AC*
/sys/class/power_supply/BAT*

On one ARM platform, for example, the names of the supplies are ac and battery.
Now the test will look for the type file and determine the type of power
supply based on this, which is proper for power supply class.

BUG=none
TEST=power_LoadTest

Signed-off-by: Benson Leung <bleung@chromium.org>

Review URL: http://codereview.chromium.org/2862044