Allow router_addr and server_addr override on command line
This allows the args in the "-a" flag to run_remote_tests to specify
the server_addr and router_addr of the config, similarly to the way
that the "--remote" argument overrides the "client_addr".  This will
be used by scripts that can find out  about the testbed config via
other means than static files.

BUG=none
TEST=Use WiFiSecMat and WiFiMatFunc with and without override

Review URL: http://codereview.chromium.org/2941013