Fix login_CryptohomeIncognitoMounted, mitigate some potential raciness.
We were running login_CryptohomeIncognitoMounted with the wrong credentials in the BVT, which is why that was failing.  As for the other, it looks like some kind of race between chrome resolving DNS names and us changing the DNS server.

We now restart chrome after we change DNS as a workaround.

BUG=5004
TEST=The BVT

Change-Id: Icf8e4bba51221d0499b8b591849443a69ec0ce20

Review URL: http://codereview.chromium.org/3018018