Return failures when other services try to auth against our mock auth server
When sync tried to use our mocked out auth service, it was sending
just an LSID instead of a SID/LSID pair.  This is legal, but not the
expected flow for normal Chrome OS auth, so we were throwing an
exception.  This seems to cause the auth server to stop so, if sync
tried to auth before the chrome os login flow completed, we were
failing to complete the chrome os login flow.  This change makes us
rebust to this, and also pulls chrome_log files back when the test
finishes, to help debug issues in the future.

BUG=5130
TEST=Run login_LoginSuccess

Change-Id: Iab175025776f41ea12c907900c368eefe97b0e64

Review URL: http://codereview.chromium.org/3013033