KVM test: Updating tests_base.cfg with cdrom_cd1
The latest cdrom patch from michael stablishes cdrom_cd1
as the first cd used by vms, but in all unattended
installs we use only cdrom, which makes it to not honor
the index set to it. Let's define cdrom_cd1 index to be 1,
and change all cdrom entries to cdrom_cd1. Also, update
the unattended install script to retrieve the cdrom from
the environment variable cdrom_cd1 instead of cdrom.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>


git-svn-id: http://test.kernel.org/svn/autotest/trunk@4735 592f7852-d20e-0410-864c-8624ca9c26a4