This is a test to verify Cryptohome behavior on TPM clear/re-own (vault must be deleted).
The test will only run on hardware with firmware that auto-enables the TPM on boot.  The test creates a cryptohome that is TPM-protected and places a file in it.  It then reboots the host, remounts that cryptohome, and checks for the presence of the file.  It then clears the TPM, reboots, and remounts the cryptohome.  Since the TPM was cleared, it expects that the cryptohome vault would be recreated, so it verifies that the file does not exist.

BUG=5317
TEST=autotest

Review URL: http://codereview.chromium.org/3085026