Correctly auto-detect 2GHz and 5GHz phy devices
Fixed already existing code to parse "iw list" so that it can correctly
detect the 5GHz and 2GHz radios on a linux router.  Also, set the
default ethernet interface to "eth0" since this is the default in our
setups and "eth1" seems strange for a default.

TEST=reran autotest with no phydevs specified
BUG=none

Review URL: http://codereview.chromium.org/3148006