Add a test for multicard reader support.
This does not test a specific reader.  It really just tests for the
existance of some removable block device that we can assume comes from
a multicard reader.

This means that USB sticks and SD cards in the internal reader should not
be present because they also provide removable block devices.

BUG=none
TEST=Automatic:  Plug in a multicard reader with media and run this test.

Change-Id: I1573d6ee340e3d2799d68cf04074d12dbe582e57

Review URL: http://codereview.chromium.org/3184017