autotest: Simplify desktopui_WindowManagerHotkeys.
This test was broken, probably as a result of the crosh
stuff that added an extra step to get to a bash shell.
Since that doesn't seem particularly useful to test anyway,
I've removed it and changed the logic that we use to verify
that the xterm window is focused.

I've also removed the volume key stuff that would've run
later in the test, since I think it's also obsolete now (the
test was checking the ALSA volume level, but I believe that
Chrome has been modified to use PulseAudio instead).

Change-Id: Ia623ae087fe77b56c40cbde6e0b6e8397e58e931

BUG=chromium-os:3098
TEST=ran it.  it fails with "Timed out waiting for logout" after all of the test-specific code has passed, but I'm pretty sure that that's just happening because I'm booting off of a slow USB stick (the logout completes soon after the timeout)

Review URL: http://codereview.chromium.org/3427008