Create implicit concept of iterations in wifitest
We are losing a lot of data because steps don't know
if they are repeated, and don't enter separate keyvals
for each iteration.  This CL addresses this by first
assessing all repeated steps in a test and assigning
iterators to each one of them.  Then as the steps are
being processed, provide a "prefix" variable that is
prepended to each keyval.  This has the handy side-effect
of the step not needing to have to repeat its method
name every time it writes out a variable.

BUG=none
TEST=run_remote_tests

Review URL: http://codereview.chromium.org/3459004